refactor: rename to dashboard-error
