+++ src/react/dashboard-app/calculation-cell/calculation.ts
@@ -12,7 +12,7 @@ import { hashString, round2Digits } from "./utils";
 import {
 	RowNotFoundError,
 	TagNotFoundError,
-} from "src/shared/dashboard-state/table-error";
+} from "src/shared/dashboard-state/dashboard-error";
 import { unixTimeToDateTimeString } from "src/shared/date/date-conversion";
 
 export const getCalculationContent = (
