+++ src/react/import-app/column-select.tsx
@@ -1,5 +1,5 @@
 import { CellType, Column, HeaderCell } from "src/shared/types";
-import { CellNotFoundError } from "src/shared/dashboard-state/table-error";
+import { CellNotFoundError } from "src/shared/dashboard-state/dashboard-error";
 import Stack from "../shared/stack";
 
 interface Props {
