+++ src/react/import-app/utils.ts
@@ -1,7 +1,7 @@
 import { DashboardState } from "src/shared/types";
 import { MARKDOWN_LIST_ITEM_REGEX } from "./constants";
 import { ImportType } from "./types";
-import { CellNotFoundError } from "src/shared/dashboard-state/table-error";
+import { CellNotFoundError } from "src/shared/dashboard-state/dashboard-error";
 import RowAddCommand from "src/shared/commands/row-add-command";
 
 export const getDisplayNameForImportType = (type: ImportType) => {
