+++ src/shared/commands/column-delete-command.ts
@@ -1,4 +1,4 @@
-import { ColumNotFoundError } from "../dashboard-state/table-error";
+import { ColumNotFoundError } from "../dashboard-state/dashboard-error";
 import {
 	Column,
 	DashboardState,
