+++ src/shared/commands/column-type-update-command.ts
@@ -1,7 +1,7 @@
 import { createTag } from "src/data/table-state-factory";
 import { unixTimeToDateString } from "../date/date-conversion";
 import { CHECKBOX_MARKDOWN_UNCHECKED } from "../dashboard-state/constants";
-import { ColumNotFoundError } from "../dashboard-state/table-error";
+import { ColumNotFoundError } from "../dashboard-state/dashboard-error";
 import DashboardStateCommand from "../dashboard-state/table-state-command";
 import {
 	BodyCell,
