+++ src/shared/commands/row-sort-command.ts
@@ -3,7 +3,7 @@ import {
 	ColumNotFoundError,
 	RowNotFoundError,
 	TagNotFoundError,
-} from "../dashboard-state/table-error";
+} from "../dashboard-state/dashboard-error";
 import DashboardStateCommand from "../dashboard-state/table-state-command";
 import {
 	BodyCell,
