+++ src/shared/dashboard-state/dashboard-error.ts
