+++ src/shared/export/table-state-to-array.tsx
@@ -6,7 +6,7 @@ import {
 	DashboardState,
 } from "../types";
 import { getCellContent } from "../cell-content";
-import { ColumNotFoundError } from "../dashboard-state/table-error";
+import { ColumNotFoundError } from "../dashboard-state/dashboard-error";
 
 const serializeHeaderCells = (cells: HeaderCell[]): string[] => {
 	return cells.map((cell) => cell.markdown);
